feat(service): 优化角色属性展示和技能解析
- 在 BattleService 中添加角色生命值展示
- 优化 KwGameApi 中的进度条方法，提高可读性- 改进 SkillController 中的技能解析逻辑，支持更多格式

+++ src/main/java/top/kloping/api/KwGameApi.java
@@ -68,7 +68,7 @@ public abstract class KwGameApi {
     }
 
     public static String getProgressBar(int current, int total, int length, String emptyChar, String filledChar) {
-        int filledLength = getFilledLength((double) current, total, length);
+        int filledLength = getFilledLength(current, total, length);
         if (total==0) filledLength = length;
         return String.valueOf(filledChar).repeat(Math.max(0, filledLength)) + String.valueOf(emptyChar).repeat(Math.max(0, length - filledLength));
     }
